# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Orange Money API",
        "description": "API for Orange Money payment system with US 2.0 compliance",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API server"
        }
    ],
    "paths": {
...
        "/v1/transactions/transfer": {
            "post": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Initiate money transfer",
                "description": "Initiate a money transfer to another account",
                "operationId": "0f5162282889e62cbecf0a964784f9a3",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "destinataire_numero",
                                    "montant"
                                ],
                                "properties": {
                                    "destinataire_numero": {
                                        "type": "string",
                                        "example": "771234567"
                                    },
                                    "montant": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 5000
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Transfert pour achat"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Transfer initiated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Transfert initié. Veuillez saisir le code de vérification."
                                        },
                                        "data": {
                                            "properties": {
                                                "transaction_id": {
                                                    "type": "integer",
                                                    "example": 123
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/v1/transactions/payment": {
            "post": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Initiate merchant payment",
                "description": "Initiate a payment to a merchant",
                "operationId": "76117ffba0a0088c51c65f6c250a527c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "marchand_id",
                                    "montant"
                                ],
                                "properties": {
                                    "marchand_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "montant": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 2500
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Paiement pour services"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Payment initiated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Paiement initié. Veuillez saisir le code de vérification."
                                        },
                                        "data": {
                                            "properties": {
                                                "transaction_id": {
                                                    "type": "integer",
                                                    "example": 124
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/v1/transactions/verify-code": {
            "post": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Verify and complete transaction",
                "description": "Verify transaction with code and complete the transaction",
                "operationId": "863638accc4ed8ed0cc45f103a6471a6",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "transaction_id",
                                    "code"
                                ],
                                "properties": {
                                    "transaction_id": {
                                        "type": "integer",
                                        "example": 123
                                    },
                                    "code": {
                                        "type": "string",
                                        "example": "123456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Transaction completed successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Transaction validée avec succès"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Transaction"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Transaction": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "reference": {
                        "type": "string",
                        "example": "TXN20241112001"
                    },
                    "type": {
                        "type": "string",
                        "enum": [
                            "transfert",
                            "paiement",
                            "depot",
                            "retrait",
                            "achat_credit"
                        ],
                        "example": "transfert"
                    },
                    "montant": {
                        "type": "number",
                        "format": "float",
                        "example": 5000
                    },
                    "frais": {
                        "type": "number",
                        "format": "float",
                        "example": 50
                    },
                    "statut": {
                        "type": "string",
                        "enum": [
                            "en_attente",
                            "validee",
                            "echouee",
                            "annulee"
                        ],
                        "example": "validee"
                    },
                    "description": {
                        "type": "string",
                        "example": "Transfert vers John Doe"
                    },
                    "date_execution": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "emetteur": {
                        "$ref": "#/components/schemas/Compte"
                    },
                    "destinataire": {
                        "$ref": "#/components/schemas/Compte"
                    },
                    "marchand": {
                        "type": "object"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format: Bearer {token}",
                "name": "Authorization",
                "in": "header"
            },
            "passport": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}